Refactored utils module into package
