+++ delphi/utils/__init__.py
@@ -1,0 +1,4 @@
+from delphi.utils.fp import *
+from delphi.utils.web import *
+from delphi.utils.shell import *
+from delphi.utils.misc import *
