+++ delphi/visualization.py
@@ -17,7 +17,6 @@ from pygraphviz import AGraph
 # Visualization
 # ==========================================================================
 
-
 @singledispatch
 def visualize():
     pass
